+++ engine/src/execution_graph/logic_controllers/LogicalFilter.cpp
@@ -20,7 +20,6 @@
 #include "distribution/primitives.h"
 #include "communication/CommunicationData.h"
 #include "utilities/CommonOperations.h"
-#include "utilities/DebuggingUtils.h"
 #include "utilities/StringUtils.h"
 #include "error.hpp"